+++ docs/Feedback.md
@@ -13,10 +13,10 @@
   - [x] when이 아닌 given
 
 ## Level 2 - 점진적인 리팩터링
-- [ ] ManualHandlerMapping, 추가된 핸들러 중 사용하지 않는 클래스는 삭제
+- [x] ManualHandlerMapping, 추가된 핸들러 중 사용하지 않는 클래스는 삭제
 - [x] 공백 제거 
-- [ ] Registry 클래스 패키지 수정
-- [ ] equals, hashcode 메서드 수정 이유 설명
+- [x] Registry 클래스 패키지 수정
+- [x] equals, hashcode 메서드 수정 이유 설명
   - 리팩터링 미션하면서 ResgisterController를 어노테이션을 사용하는 방식으로 수정함.
   - 그 과정에서 입력한 Request에 따른 Handler가 검색되지 않는 에러가 발생함.
   - 해당 과정에서 문제점이 equals 메서드인지 의심이 되어 수정해보았음.
