+++ mvc/src/main/java/nextstep/mvc/HandlerAdapterRegistry.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.controller;
+package nextstep.mvc;
 
 import java.util.ArrayList;
 import java.util.List;
